move tests fns outside of top level index
